requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
whitenoise==6.6.0
gunicorn==21.2.0
Pillow==10.2.0 